ml/requirements.txt
tensorflow~=2.20
numpy
pillow
